+++ eventmesh-runtime/src/main/java/org/apache/eventmesh/runtime/core/retry/RetryContext.java
@@ -17,8 +17,6 @@
 
 package org.apache.eventmesh.runtime.core.retry;
 
-import org.apache.eventmesh.runtime.core.protocol.DelayRetryable;
-
 import java.util.concurrent.Delayed;
 import java.util.concurrent.TimeUnit;
